llm_isolated_service/EduNavigator/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pypdf==5.1.0
faiss-cpu==1.12.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.15
sentence-transformers==3.2.1
numpy==1.24.3
scikit-learn==1.5.2
google-generativeai==0.8.3
orjson==3.10.7
aiofiles==24.1.0
jinja2==3.1.4

